Actualizacion al 27/10

+++ dj/heroesApp/heroe/views.py
@@ -1,13 +1,17 @@
 # Rest Imports
 from rest_framework.views import APIView
+from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework import status
 
 # Models Imports
 from heroe.models import Hero
 
-# Serializer import
+# Serializer Imports
 from heroe.serializer import HeroSerializer
+
+# Helpers Imports
+from heroe.helpers.heroeErrors import hayHeroe
 
 # Create your views here.
 
@@ -24,13 +28,22 @@ class HeroApiView(APIView):
             status=status.HTTP_200_OK
         )
     
+@api_view(["GET"])
+def hero_api_view(request):
+    heroes = Hero.objects.all()
+    heroes_serializer = HeroSerializer(heroes, many = True)
+
+    return Response(
+        data=heroes_serializer.data,
+        status=status.HTTP_200_OK
+    )
 
 class CrearHeroApiView(APIView):
 
     def post(self, request):
         """Crear un nuevo registro/heroe"""
         print("Segundo POST")
-        serializer = HeroSerializer(data = request.data)
+        serializer = HeroSerializer(data = request.data, many = True)
 
         if serializer.is_valid():
             serializer.save()
@@ -50,12 +63,86 @@ class CrearHeroApiView(APIView):
 
 
 class HeroDetalleApiView(APIView):
-
+    
     def get(self, request, pk):
         """Nos retorna mas info de un heroe en particular"""
 
-        heroe = Hero.objects.get(pk=pk)
+        try:
+            heroe = Hero.objects.get(pk=pk)
 
+            heroe_serializer = HeroSerializer(heroe)
+
+            return Response(
+                data=heroe_serializer.data,
+                status=status.HTTP_200_OK
+            )
+        except:        
+            data = {
+                "message":"El heroe no fue encontrado"
+            }
+
+            return Response(
+                data=data,
+                status = status.HTTP_400_BAD_REQUEST
+            )
+
+    def put(self, request, pk):
+        """Nos modifica un heroe en particular"""
+
+        heroe = hayHeroe(pk)
+        if heroe[0]:
+
+            heroe_serializer = HeroSerializer((heroe)[1], data=request.data)
+
+            if heroe_serializer.is_valid():
+                heroe_serializer.save()
+            
+                data = {
+                    "message":"El heroe fue modificado de forma correcta"
+                }
+
+                return Response(
+                    data=data,
+                    status=status.HTTP_200_OK
+                )
+            
+        return Response(
+            data = heroe_serializer.errors,
+            status = status.HTTP_400_BAD_REQUEST
+        )
+
+
+    def delete(self, request, pk):
+        """Elimina un heroe"""
+
+        heroe = Hero.objects.get(pk=pk)
+        heroe.delete()
+
+        data = {
+            "message":"El heroe fue eliminado de forma correcta"
+        }
+
+        return Response(
+            data=data,
+            status=status.HTTP_200_OK
+        )
+
+
+@api_view(["GET", "PUT", "DELETE"])
+def hero_detail_api_view(request, pk):
+
+    try:
+        heroe = Hero.objects.get(pk=pk) 
+    
+    except:
+        return Response(
+        {"Mensaje":"Heroe No Encontrado"},
+        status=status.HTTP_400_BAD_REQUEST
+        )
+    
+    # Detail
+    if request.method == "GET":
+    
         heroe_serializer = HeroSerializer(heroe)
 
         return Response(
@@ -63,10 +150,9 @@ class HeroDetalleApiView(APIView):
             status=status.HTTP_200_OK
         )
 
-    def put(self, request, pk):
-        """Nos modifica un heroe en particular"""
 
-        heroe = Hero.objects.get(pk=pk)
+    # Actualizar
+    elif request.method == "PUT":
 
         heroe_serializer = HeroSerializer(heroe, data=request.data)
 
@@ -81,21 +167,17 @@ class HeroDetalleApiView(APIView):
                 data=data,
                 status=status.HTTP_200_OK
             )
-        
-        return Response(
-            data = heroe_serializer.errors,
-            status = status.HTTP_400_BAD_REQUEST
-        )
+            
+    
+    # Borrar
+    elif request.method == "DELETE":
 
-    def delete(self, request, pk):
-        """Elimina un heroe"""
-
-        heroe = Hero.objects.get(pk=pk)
         heroe.delete()
 
         data = {
             "message":"El heroe fue eliminado de forma correcta"
         }
+
         return Response(
             data=data,
             status=status.HTTP_200_OK
